requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
chainlit>=1.1.0
langchain>=0.2.0
langgraph>=0.0.71
langchain-openai>=0.1.21
httpx>=0.27.0
pydantic-settings>=2.2.1
openpyxl>=3.1.2
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-httpx>=0.46b0
